requirements.txt
click==8.0.1
Flask==2.0.1
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
numpy==1.21.1
pandas==1.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
xlwt==1.3.0
